Route refactor and healthcheck improvements
